# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cats on the Street API",
    "version": "1.0.0",
    "description": "Mock API for tracking cats on the streets of London."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/cats": {
      "get": {
        "summary": "Get all cat names",
        "responses": {
          "200": {
            "description": "List of cat names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new cat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "description", "location"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Snowball"
                  },
                  "description": {
                    "type": "string",
                    "example": "White cat in Hackney"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New cat created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 6
                    },
                    "name": {
                      "type": "string",
                      "example": "Snowball"
                    },
                    "description": {
                      "type": "string",
                      "example": "White cat in Hackney"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cats/{id}": {
      "get": {
        "summary": "Get a cat by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cat object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cat"
                }
              }
            }
          },
          "404": {
            "description": "Cat not found"
          }
        }
      },
      "put": {
        "summary": "Update a cat's description",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "example": "Updated description"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated cat object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cat"
                }
              }
            }
          },
          "404": {
            "description": "Cat not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a cat by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cat deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Cat deleted"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Cat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Whiskers"
          },
          "description": {
            "type": "string",
            "example": "Siamese grey in Stockwell"
          }
        }
      }
    }
  }
}